Updated FNF condition
Updated the code to grab another version of allWords.csv from GitHub instead of halting the program on an error if both allWords.csv and wordleWords.csv are missing. This also allows users to only have to delete these two files if they want to update their word lists to be the most up to date and improved version.